requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
trafilatura==1.6.4
newspaper3k==0.2.8
PyPDF2==3.0.1
Pillow==10.1.0
opencv-python==4.8.1.78
pytube==15.0.0
chromadb==0.4.18
openai==1.3.7
voyageai==0.2.1
ollama==0.1.7
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
markdown==3.5.1
jinja2==3.1.2
asyncio-throttle==1.0.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
starlette==0.27.0
fastapi-mail==1.4.1
PyPDF2
python-jose[cryptography]
passlib[bcrypt]
replicate
bcrypt==4.0.1